opencv3: disable --with-quicktime for newer development tools
Since https://github.com/opencv/opencv/pull/7159,
opencv and opencv3 use AVFoundation instead of quicktime.
The same change as in this commit was done for the opencv
formula, so this brings both opencv and opencv3 in line.

The AVFoundation is ON by default on mac, so video capabilities
are still guaranteed.

Closes #5923.

Signed-off-by: Michka Popoff <michkapopoff@gmail.com>