scalapack: update 2.0.2_8 bottle for Linuxbrew.
The test is taking more than 10 minutes to respond,
making the Linux CI fail. It is disabled to be
able to build bottles.
Also, remove the head option, because the audit fails
on a broken https certifcate for the head URL.
The homebrew statistics show that the head builds are
really few (8 in around 6 months). If somebody needs
the head build, they will need to contact the
sysadmins and fix the certificate.
The audit prevents bottles to be built.
Also fixes:
* W: 44: col 32: Use parentheses in the method call to avoid confusion about precedence.
* C: 55: col 15: Use ``mpirun -np 4 ./xsscaex | grep 'INFO code' | awk '{print $NF}'`.to_i.zero?` instead of ``mpirun -np 4 ./xsscaex | grep 'INFO code' | awk '{print $NF}'`.to_i == 0`.
* C: 3: col 9: Description should start with a capital letter
* C: 53: col 14: Don't use parentheses around a method call.
* C: 55: col 14: Don't use parentheses around a method call.
* C: 57: col 14: Don't use parentheses around a method call.
* C: 59: col 14: Don't use parentheses around a method call.
Closes #5928.
Signed-off-by: Michka Popoff <michkapopoff@gmail.com>